STOR-5120 - Query builder now supplies the select statement to gorm
By default gorm uses "select *", unless the select statement is
explicitly supplied.
Query builder uses the model to construct the select statement which
allows selecting only fields that are in the model.
This commit also added an optional 'query' tag - 'detailed'
If the Query builder is configured in 'OnlySelectNonDetailedFields'
mode, it will not add the fields tagged with 'detailed' to the select
statement thus allowing the user to avoid selecting costly fields (such
as TEXT fields) if not required